requirements.txt
amqp==2.5.1
Babel==2.7.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
coverage==4.5.4
defusedxml==0.6.0
Django==2.2.4
django-filter==2.2.0
django-nested-inline==0.3.7
django-phonenumber-field==3.0.1
django-rest-passwordreset==1.0.0
django-rest-recaptcha==1.0.2
djangorestframework==3.10.2
djangorestframework-xml==1.4.0
djangorestframework-yaml==1.0.3
gevent==1.4.0
greenlet==0.4.15
idna==2.8
importlib-metadata==0.20
kombu==4.6.4
more-itertools==7.2.0
phonenumbers==8.10.17
pycparser==2.19
pytz==2019.2
PyYAML==5.1.2
redis==3.3.8
requests==2.22.0
sqlparse==0.3.0
ujson==1.35
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
wincertstore==0.2
zipp==0.6.0
